// src/lib/queries/marketplace/marketplace.gen.ts
export enum MarketplaceKind {
  unknown = 'unknown',
  sequence_marketplace_v1 = 'sequence_marketplace_v1',
  sequence_marketplace_v2 = 'sequence_marketplace_v2',
  opensea = 'opensea',
  magic_eden = 'magic_eden',
  mintify = 'mintify',
  looks_rare = 'looks_rare',
  x2y2 = 'x2y2',
  sudo_swap = 'sudo_swap',
  coinbase = 'coinbase',
  rarible = 'rarible',
  nftx = 'nftx',
  foundation = 'foundation',
  manifold = 'manifold',
  zora = 'zora',
  blur = 'blur',
  super_rare = 'super_rare',
  okx = 'okx',
  element = 'element',
  aqua_xyz = 'aqua_xyz',
  auranft_co = 'auranft_co'
}
// ...
export enum OrderSide {
  unknown = 'unknown',
  listing = 'listing',
  offer = 'offer'
}

export enum OrderStatus {
  unknown = 'unknown',
  active = 'active',
  inactive = 'inactive',
  expired = 'expired',
  cancelled = 'cancelled',
  filled = 'filled'
}
// ...
export interface Page {
  page: number
  pageSize: number
  more?: boolean
  sort?: Array<SortBy>
}

export interface SortBy {
  column: string
  order: SortOrder
}

export interface Filter {
  text?: string
  properties?: Array<PropertyFilter>
}

export interface PropertyFilter {
  name: string
  type: PropertyType
  min?: number
  max?: number
  values?: Array<any>
}

export interface CollectiblesFilter {
  includeEmpty: boolean
  searchText?: string
  properties?: Array<PropertyFilter>
  marketplaces?: Array<MarketplaceKind>
  inAccounts?: Array<string>
  notInAccounts?: Array<string>
  ordersCreatedBy?: Array<string>
  ordersNotCreatedBy?: Array<string>
}

export interface Order {
  orderId: string
  marketplace: MarketplaceKind
  side: OrderSide
  status: OrderStatus
  chainId: number
  collectionContractAddress: string
  tokenId: string
  createdBy: string
  priceAmount: string
  priceAmountFormatted: string
  priceAmountNet: string
  priceAmountNetFormatted: string
  priceCurrencyAddress: string
  priceDecimals: number
  priceUSD: number
  quantityInitial: string
  quantityInitialFormatted: string
  quantityRemaining: string
  quantityRemainingFormatted: string
  quantityAvailable: string
  quantityAvailableFormatted: string
  quantityDecimals: number
  feeBps: number
  feeBreakdown: Array<FeeBreakdown>
  validFrom: string
  validUntil: string
  orderCreatedAt?: string
  orderUpdatedAt?: string
  createdAt: string
  updatedAt: string
  deletedAt?: string
}

export interface FeeBreakdown {
  kind: string
  recipientAddress: string
  bps: number
}

export interface CollectibleOrder {
  metadata: TokenMetadata
  order?: Order
}

export interface OrderFilter {
  createdBy?: Array<string>
  marketplace?: Array<MarketplaceKind>
  currencies?: Array<string>
}
// ...
export interface CheckoutOptionsMarketplaceOrder {
  contractAddress: string
  orderId: string
  marketplace: MarketplaceKind
}
// ...
export interface GetCollectibleHighestOfferArgs {
  contractAddress: string
  tokenId: string
  filter?: OrderFilter
}

export interface GetCollectibleHighestOfferReturn {
  order: Order  
}
export interface GetCollectibleLowestListingArgs {
  contractAddress: string
  tokenId: string
  filter?: OrderFilter
}
// ...
export interface ListCollectiblesWithLowestListingArgs {
  contractAddress: string
  filter?: CollectiblesFilter
  page?: Page
}

export interface ListCollectiblesWithLowestListingReturn {
  collectibles: Array<CollectibleOrder>
  page?: Page  
}
export interface ListCollectiblesWithHighestOfferArgs {
  contractAddress: string
  filter?: CollectiblesFilter
  page?: Page
}

export interface ListCollectiblesWithHighestOfferReturn {
  collectibles: Array<CollectibleOrder>
  page?: Page  
}
export interface SyncOrderArgs {
  order: Order
}

export interface SyncOrderReturn {  
}
export interface SyncOrdersArgs {
  orders: Array<Order>
}

export interface SyncOrdersReturn {  
}
export interface GetOrdersArgs {
  input: Array<GetOrdersInput>
  page?: Page
}

export interface GetOrdersReturn {
  orders: Array<Order>
  page?: Page  
}
export interface CheckoutOptionsMarketplaceArgs {
  wallet: string
  orders: Array<CheckoutOptionsMarketplaceOrder>
  additionalFee: number
}